requirements.txt
flask
pandas
matplotlib
wordcloud
networkx
python-dotenv
